shopping cart added + refactoring + design frontend

+++ jewelry-server/src/main/java/ro/ausy/jewelry/server/dao/ProductDao.java
@@ -13,7 +13,7 @@ import ro.ausy.jewelry.server.entity.Product;
  */
 public interface ProductDao {
 	
-	ProductDTO getProductById(long productId);
+	ProductDTO getProductById(int productId);
 	
 	List<ProductDTO> getAllProduct();
