+++ jewelry-server/src/main/java/ro/ausy/jewelry/server/dao/impl/ProductDaoImpl.java
@@ -84,7 +84,7 @@ public class ProductDaoImpl implements ProductDao {
 	}
 	
 	@Override
-	public ProductDTO getProductById(long productId) {
+	public ProductDTO getProductById(int productId) {
 		LOGGER.info("ProductDaoImpl getProductById(Long id) has been called");
 		
 		ProductDTO productDTO = new ProductDTO();
